// Dockerfile/src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Sublog = {
  __typename?: 'Sublog';
  id: Maybe<Scalars['String']>;
  createdAt: Maybe<Scalars['String']>;
  fileName: Maybe<Scalars['String']>;
  category: Maybe<Scalars['String']>;
  media: Maybe<Scalars['String']>;
  title: Maybe<Scalars['String']>;
  eyeCatchURL: Maybe<Scalars['String']>;
  tag: Maybe<Array<Maybe<Scalars['String']>>>;
  updatedAt: Maybe<Scalars['String']>;
};